pyproject.toml
[tool.poetry]
name = "djangofrontend"
version = "0.1.0"
description = "nothing"
authors = ["Nauman khan gori <khangori850@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.2"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
